Changed user repository to implement jpa repo, Added h2 embedded db

+++ src/main/java/org/studnia/chatapp/services/user/UserService.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface UserService {
     List<UserDTO> getAllUsers();
 
-    UserDTO getSingleUser(int id);
+    UserDTO getSingleUser(long id);
 
     void registerUser(UserDTO newUser);
 
-    void deleteUser(int id);
+    void deleteUser(long id);
 }
